OSG-RCC relay 1: Add RCC jobs submitters (#278)
* List all files including per-chunk results

* Add options of OSG-RCC relay

* Use `WORKFLOW_DIR` to indicate the location of workflow and jobs

* Assert admix + gfal2 availability

* Add the per-chunk processing in different folders

+++ outsource/workflow/combine-wrapper.sh
@@ -9,18 +9,25 @@ rucio_upload=$4
 rundb_update=$5
 stage=$6
 remove_heavy=$7
-tar_filename=$8
+input_path=$8
+output_path=$9
+staging_dir=${10}
+tar_filename=${11}
 args=( "$@" )
-chunks=${args[@]:8}
+chunks=${args[@]:11}
 
 echo $@
 echo $*
+echo
 
+echo "Where am i:"
+echo $PWD
+echo
+
+# Needed by utilix DB
 export HOME=$PWD
 
-input_path="input"
 mkdir -p $input_path
-output_path="output"
 mkdir -p $output_path
 
 extraflags=""
@@ -52,8 +59,13 @@ if [ -e /image-build-info.txt ]; then
     echo
 fi
 
-# Installing customized packages
-. install.sh strax straxen cutax utilix admix outsource
+unset http_proxy
+export XENON_CONFIG=$PWD/.xenon_config
+
+if [ -f install.sh ]; then
+    # Installing customized packages
+    . install.sh strax straxen cutax utilix admix outsource
+fi
 
 echo "Current dir is $PWD. Here's whats inside:"
 ls -lah .
@@ -63,8 +75,13 @@ if [ "X$rucio_upload" = "Xtrue" ]; then
     export RUCIO_ACCOUNT=production
 fi
 
-unset http_proxy
-export XENON_CONFIG=$PWD/.xenon_config
+echo "PEGASUS Stuff:"
+env | grep PEGASUS
+echo
+
+echo "XENON Stuff:"
+env | grep XENON
+echo
 
 echo "RUCIO/X509 Stuff:"
 env | grep RUCIO
@@ -87,11 +104,11 @@ echo "What is in the input directory:"
 ls -lah $input_path
 
 echo
-echo "Total amount of data before combine: "`du -s --si $input_path | cut -f1`
+echo "Total amount of data before combining: "`du -s --si $input_path | cut -f1`
 echo
 
 echo "Combining:"
-time python3 combine.py $run_id --context $context --xedocs_version $xedocs_version --input_path $input_path --output_path $output_path $chunksarg $extraflags
+time python3 combine.py $run_id --context $context --xedocs_version $xedocs_version --input_path $input_path --output_path $output_path --staging_dir $staging_dir $chunksarg $extraflags
 
 echo
 echo "Moving auxiliary files to output directory"
@@ -104,8 +121,11 @@ echo
 echo "Total amount of data in $input_path before removing: "`du -s --si $input_path | cut -f1`
 echo
 
-echo "Removing inputs directory:"
-rm -r $input_path
+# There will not be storage pressure if no tarball is produced
+if [ $tar_filename != "X" ]; then
+    echo "Removing inputs directory:"
+    rm -r $input_path
+fi
 
 echo "Here is what is in the output directory after combining:"
 ls -lah $output_path
@@ -114,9 +134,11 @@ echo
 echo "Total amount of data in $output_path before tarballing: "`du -s --si $output_path | cut -f1`
 echo
 
-echo "We are tarballing the output directory and removing it:"
-tar czfv $tar_filename $output_path
-# tar czfv $tar_filename $output_path --remove-files
+if [ $tar_filename != "X" ]; then
+    echo "We are tarballing the output directory and removing it:"
+    tar czfv $tar_filename $output_path
+    # tar czfv $tar_filename $output_path --remove-files
+fi
 
 echo
 echo "Job is done. Here is the contents of the directory now:"
